refactor(sonarr_compat): split 1288-line file into 5-module subdir
  sonarr_compat/
    mod.rs      47   router glue: auth middleware + pub re-exports
    types.rs   166   SonarrSeries + 5 nested types + 5 request DTOs
    system.rs   98   system_status, quality_profiles, root_folders,
                      language_profiles, list_tags, list_download_clients,
                      create_tag (the small-surface handlers)
    series.rs  583   series_lookup, list_series, get_series, add_series
                      (394-line squashed-merge + regrab), update_series,
                      execute_command
    helpers.rs 445   lookup_by_external_id, fetch_anime_detail,
                      build_sonarr_series_from_search / _from_tracked,
                      map_status, build_stub_series

Visibility bumps applied during extraction (all caught by rustc):
  - 6 helper fns → pub(super) (called by series.rs from a sibling)
  - RootFolder + LanguageProfile fields → pub (cross-module construction)

Path fix: `super::library::search::auto_search_series` →
`super::super::library::search::auto_search_series` (one extra super
to hop out of the new sonarr_compat/ subdir).

External handler paths unchanged — main.rs still calls
sonarr_compat::system_status, sonarr_compat::add_series, etc. via
pub use re-exports in mod.rs.

cargo fmt / clippy -D warnings / build / test all green (601 tests).
